Created WarningNote Post Endpoint
- New infrastructure objects:
-- Warning Note Set
-- Warning Note Case Note

- New Boundary objects:
-- CreateWarningNoteRequest
-- CreateWarningNoteResponse

- CreateWarningNote method created in DatabaseGateway
- WarningNoteUseCase object created with ExecutePost method
- warningnotes route created in controller
- CreateWarningNote method created in SCCVAPIController

+++ SocialCareCaseViewerApi/V1/Boundary/Response/CreateWarningNoteResponse.cs
@@ -3,6 +3,6 @@ namespace SocialCareCaseViewerApi.V1.Boundary.Response
     public class CreateWarningNoteResponse
     {
         public string CaseNoteId { get; set; }
-        public int WarningNoteId { get; set; }
+        public long WarningNoteId { get; set; }
     }
 }
